Refactor projects page to use shared utilities
- Create `pages/projekte/project-utils.js` to centralize GitHub API and URL logic.
- Update `pages/projekte/projects-data.js` to use shared API functions while keeping caching.
- Update `pages/projekte/projekte-app.js` to use shared URL helpers.
- Update `pages/projekte/update-apps-config.js` to use shared logic for maintenance.
