+++ pages/projekte/update-apps-config.js
@@ -5,77 +5,22 @@
  * Run this manually when you add new apps to keep the local config in sync
  */
 
-const GITHUB_CONFIG = {
-  owner: 'aKs030',
-  repo: 'Webgame',
-  branch: 'main',
-  appsPath: 'apps',
-  apiBase: 'https://api.github.com',
-  rawBase: 'https://raw.githubusercontent.com',
-};
-
-async function fetchGitHubContents(path = '') {
-  const url = `${GITHUB_CONFIG.apiBase}/repos/${GITHUB_CONFIG.owner}/${GITHUB_CONFIG.repo}/contents/${path}`;
-
-  try {
-    const response = await fetch(url, {
-      headers: {
-        Accept: 'application/vnd.github.v3+json',
-        'User-Agent': 'Apps-Config-Updater/1.0',
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error(`GitHub API error: ${response.status}`);
-    }
-
-    return await response.json();
-  } catch (error) {
-    console.error('Failed to fetch GitHub contents:', error);
-    return [];
-  }
-}
-
-async function fetchProjectMetadata(projectPath) {
-  const metadataUrl = `${GITHUB_CONFIG.rawBase}/${GITHUB_CONFIG.owner}/${GITHUB_CONFIG.repo}/${GITHUB_CONFIG.branch}/${projectPath}/package.json`;
-
-  try {
-    const response = await fetch(metadataUrl);
-    if (response.ok) {
-      const packageData = await response.json();
-      return {
-        name: projectPath.split('/').pop(),
-        title: packageData.name || projectPath.split('/').pop(),
-        description: packageData.description || 'Ein interaktives Web-Projekt',
-        category: packageData.category || 'app',
-        tags: packageData.keywords || ['JavaScript'],
-        version: packageData.version || '1.0.0',
-      };
-    }
-  } catch (error) {
-    console.warn(`Could not fetch metadata for ${projectPath}:`, error);
-  }
-
-  // Default metadata
-  return {
-    name: projectPath.split('/').pop(),
-    title: projectPath
-      .split('/')
-      .pop()
-      .replace(/-/g, ' ')
-      .replace(/\b\w/g, (l) => l.toUpperCase()),
-    description: 'Ein interaktives Web-Projekt',
-    category: 'app',
-    tags: ['JavaScript'],
-    version: '1.0.0',
-  };
-}
+import { GITHUB_CONFIG } from './github-config.js';
+import {
+  fetchGitHubContents,
+  fetchProjectMetadata,
+} from './project-utils.js';
+import { fileURLToPath } from 'url';
 
 async function updateAppsConfig() {
   console.log('🚀 Updating apps config from GitHub...');
 
   try {
     const contents = await fetchGitHubContents(GITHUB_CONFIG.appsPath);
+    if (!contents || !Array.isArray(contents)) {
+      throw new Error('Invalid contents received from GitHub');
+    }
+
     const directories = contents.filter((item) => item.type === 'dir');
 
     console.log(`📁 Found ${directories.length} directories`);
@@ -93,33 +38,41 @@ async function updateAppsConfig() {
       }
 
       const metadata = await fetchProjectMetadata(projectPath);
-      apps.push(metadata);
+
+      // We add 'name' as the directory name, which is required for local config fallback
+      apps.push({
+        name: dir.name,
+        ...metadata,
+      });
 
       console.log(`✅ Added: ${metadata.title}`);
     }
 
+    // Clean up 'raw' property if it exists, as we don't need it in the config file
+    const cleanApps = apps.map(({ raw, ...app }) => app);
+
     const config = {
-      apps: apps,
+      apps: cleanApps,
       lastUpdated: new Date().toISOString(),
       source: 'github-api',
     };
 
     // In a real Node.js environment, you would write to file:
-    // const fs = require('fs');
-    // fs.writeFileSync('./apps-config.json', JSON.stringify(config, null, 2));
+    // const fs = await import('fs');
+    // fs.writeFileSync('./pages/projekte/apps-config.json', JSON.stringify(config, null, 2));
 
     console.log('📝 Updated config:');
     console.log(JSON.stringify(config, null, 2));
 
     console.log('🎉 Apps config updated successfully!');
-    console.log('💡 Copy the JSON above to apps-config.json');
+    console.log('💡 Copy the JSON above to pages/projekte/apps-config.json');
   } catch (error) {
     console.error('❌ Failed to update apps config:', error);
   }
 }
 
 // Run if called directly
-if (typeof window === 'undefined') {
+if (process.argv[1] === fileURLToPath(import.meta.url)) {
   updateAppsConfig();
 }
